pyproject.toml
[project]
name = "introductiontolangchain"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
]
